# experiments/st-johns-weather-map/contracts/source-api.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IFSItem": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "grid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IFSGrid"
              },
              {
                "type": "null"
              }
            ]
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "level": {
            "title": "Level",
            "type": "integer"
          },
          "member": {
            "title": "Member",
            "type": "string"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          }
        },
        "required": [
          "index",
          "product",
          "field",
          "level",
          "run_id",
          "time",
          "member"
        ],
        "title": "IFSItem",
        "type": "object"
      },
      "IFSPage": {
        "additionalProperties": false,
        "properties": {
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_bytes": {
            "title": "Input Bytes",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/IFSItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "receipt_manifest": {
            "title": "Receipt Manifest",
            "type": "string"
          },
          "record_acquisitions": {
            "title": "Record Acquisitions",
            "type": "integer"
          },
          "selected_at": {
            "format": "date-time",
            "title": "Selected At",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "selected_at",
          "expires_at",
          "completed",
          "total",
          "input_bytes",
          "record_acquisitions",
          "items",
          "next_cursor",
          "receipt_manifest"
        ],
        "title": "IFSPage",
        "type": "object"
      },
      "IFSReceipts": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "receipts": {
            "items": {
              "$ref": "#/components/schemas/SourceTransferReceipt"
            },
            "maxItems": 12,
            "title": "Receipts",
            "type": "array"
          }
        },
        "required": [
          "id",
          "offset",
          "receipts",
          "next_offset"
        ],
        "title": "IFSReceipts",
        "type": "object"
      },
...
      "SourceTransferReceipt": {
        "additionalProperties": false,
        "properties": {
          "byte_size": {
            "minimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "effective_url": {
            "maxLength": 2048,
            "title": "Effective Url",
            "type": "string"
          },
          "http_status": {
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Http Status",
            "type": "integer"
          },
          "request_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 16,
            "title": "Request Headers",
            "type": "object"
          },
          "response_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 16,
            "title": "Response Headers",
            "type": "object"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "url": {
            "maxLength": 2048,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "effective_url",
          "http_status",
          "request_headers",
          "response_headers",
          "byte_size",
          "sha256",
          "completed_at"
        ],
        "title": "SourceTransferReceipt",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Astraeus experimental source delivery",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/experiments/weather/v0/ifs/selections/{id}": {
      "delete": {
        "operationId": "cancel_api_experiments_weather_v0_ifs_selections__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel"
      },
      "get": {
        "operationId": "page_api_experiments_weather_v0_ifs_selections__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Cursor",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IFSPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Page"
      }
    },
...
    "/api/experiments/weather/v0/ifs/selections/{id}/receipts": {
      "get": {
        "operationId": "receipts_api_experiments_weather_v0_ifs_selections__id__receipts_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IFSReceipts"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipts"
      }
    },
...
  }
}